add voxpupuli-rubocop

+++ lib/librarian/puppet/source/forge/repo.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'json'
 require 'open-uri'
 require 'librarian/puppet/util'
@@ -12,11 +14,12 @@ module Librarian
 
           def versions
             return @versions if @versions
+
             @versions = get_versions
             if @versions.empty?
               info { "No versions found for module #{name}" }
             else
-              debug { "  Module #{name} found versions: #{@versions.join(", ")}" }
+              debug { "  Module #{name} found versions: #{@versions.join(', ')}" }
             end
             @versions
           end
@@ -49,24 +52,19 @@ module Librarian
               raise Error, "Could not find a local copy of #{name} at #{version}."
             end
 
-            if environment.vendor?
-              vendor_cache(name, version) unless vendored?(name, version)
-            end
+            vendor_cache(name, version) if environment.vendor? && !vendored?(name, version)
 
             cache_version_unpacked! version
 
-            if install_path.exist? && rsync? != true
-              install_path.rmtree
-            end
+            install_path.rmtree if install_path.exist? && rsync? != true
 
             unpacked_path = version_unpacked_cache_path(version).join(module_name(name))
 
             unless unpacked_path.exist?
               raise Error, "#{unpacked_path} does not exist, something went wrong. Try removing it manually"
-            else
-              cp_r(unpacked_path, install_path)
             end
 
+            cp_r(unpacked_path, install_path)
           end
 
           def cache_version_unpacked!(version)
@@ -79,9 +77,13 @@ module Librarian
 
             module_repository = source.uri.to_s
 
-            command = %W{puppet module install --version #{version} --target-dir}
-            command.push(*[path.to_s, "--module_repository", module_repository, "--modulepath", path.to_s, "--module_working_dir", path.to_s, "--ignore-dependencies", target])
-            debug { "Executing puppet module install for #{name} #{version}: #{command.join(" ").gsub(module_repository, source.to_s)}" }
+            command = %W[puppet module install --version #{version} --target-dir]
+            command.push(path.to_s, '--module_repository', module_repository, '--modulepath', path.to_s,
+                         '--module_working_dir', path.to_s, '--ignore-dependencies', target)
+            debug do
+              "Executing puppet module install for #{name} #{version}: #{command.join(' ').gsub(module_repository,
+                                                                                                source.to_s)}"
+            end
 
             begin
               Librarian::Posix.run!(command)
@@ -89,24 +91,24 @@ module Librarian
               # Rollback the directory if the puppet module had an error
               begin
                 path.unlink
-              rescue => u
+              rescue StandardError => u
                 debug("Unable to rollback path #{path}: #{u}")
               end
-              tar = Dir[File.join(path.to_s, "**/*.tar.gz")]
-              msg = ""
+              tar = Dir[File.join(path.to_s, '**/*.tar.gz')]
+              msg = ''
               if e.message =~ /Unexpected EOF in archive/ and !tar.empty?
                 file = tar.first
                 msg = " (looks like an incomplete download of #{file})"
               end
-              raise Error, "Error executing puppet module install#{msg}. Check that this command succeeds:\n#{command.join(" ")}\nError:\n#{e.message}"
+              raise Error,
+                    "Error executing puppet module install#{msg}. Check that this command succeeds:\n#{command.join(' ')}\nError:\n#{e.message}"
             end
-
           end
 
           def vendor_cache(name, version)
             url = url(name, version)
             path = vendored_path(name, version).to_s
-            debug { "Downloading #{url} into #{path}"}
+            debug { "Downloading #{url} into #{path}" }
             environment.vendor!
             File.open(path, 'wb') do |f|
               URI.open(url, 'rb') do |input|
@@ -114,7 +116,6 @@ module Librarian
               end
             end
           end
-
         end
       end
     end
